chore(client): reorganize client folder
