+++ lib/utils/addEntries.js
@@ -36,7 +36,7 @@ function addEntries(config, options, server) {
   const sockPort = options.sockPort ? `&sockPort=${options.sockPort}` : '';
   /** @type {string} */
   const clientEntry = `${require.resolve(
-    '../../client/'
+    '../../client/default/'
   )}?${domain}${sockHost}${sockPath}${sockPort}`;
 
   /** @type {(string[] | string)} */
