+++ lib/utils/routes.js
@@ -13,13 +13,13 @@ function routes(server) {
   app.get('/__webpack_dev_server__/sockjs.bundle.js', (req, res) => {
     res.setHeader('Content-Type', 'application/javascript');
 
-    createReadStream(join(clientBasePath, 'sockjs.bundle.js')).pipe(res);
+    createReadStream(join(clientBasePath, 'sockjs/sockjs.bundle.js')).pipe(res);
   });
 
   app.get('/webpack-dev-server.js', (req, res) => {
     res.setHeader('Content-Type', 'application/javascript');
 
-    createReadStream(join(clientBasePath, 'index.bundle.js')).pipe(res);
+    createReadStream(join(clientBasePath, 'default/index.bundle.js')).pipe(res);
   });
 
   app.get('/invalidate', (_req, res) => {
