+++ test/client/socket-helper.test.js
@@ -8,7 +8,7 @@ describe('socket', () => {
 
   it('should default to WebsocketClient when no __webpack_dev_server_client__ set', () => {
     jest.mock('../../client/clients/WebsocketClient');
-    const socket = require('../../client/socket');
+    const socket = require('../../client/default/socket');
     const WebsocketClient = require('../../client/clients/WebsocketClient');
 
     const mockHandler = jest.fn();
@@ -36,7 +36,7 @@ describe('socket', () => {
 
   it('should use __webpack_dev_server_client__ when set', () => {
     jest.mock('../../client/clients/WebsocketClient');
-    const socket = require('../../client/socket');
+    const socket = require('../../client/default/socket');
     global.__webpack_dev_server_client__ = require('../../client/clients/WebsocketClient');
 
     const mockHandler = jest.fn();
